refactor: finished up tracker refactor
