+++ lib/features/tracker/data/models/current_workout_model.dart
@@ -1,5 +1,6 @@
 import 'package:flex_workout_mobile/features/exercise/data/models/exercise_model.dart';
 import 'package:flex_workout_mobile/features/exercise/data/models/muscle_group_model.dart';
+import 'package:flex_workout_mobile/features/tracker/data/db/tracked_workout_entity.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/tracked_workout_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
@@ -27,6 +28,12 @@ class CurrentWorkoutSection with _$CurrentWorkoutSection {
   }) = _CurrentWorkoutSection;
 }
 
+extension ConvertCurrentWorkoutSection on CurrentWorkoutSection {
+  WorkoutSection toEntity() => WorkoutSection(
+        title: title,
+      )..organizers.addAll(organizers.map((e) => e.toEntity()).toList());
+}
+
 @freezed
 class CurrentWorkoutOrganizer with _$CurrentWorkoutOrganizer {
   const factory CurrentWorkoutOrganizer({
@@ -35,6 +42,14 @@ class CurrentWorkoutOrganizer with _$CurrentWorkoutOrganizer {
     CurrentWorkoutSetType? defaultSet,
     @Default([]) List<CurrentWorkoutSetType> superSet,
   }) = _CurrentWorkoutOrganizer;
+}
+
+extension ConvertCurrentWorkoutOrganizer on CurrentWorkoutOrganizer {
+  SetOrganizer toEntity() => SetOrganizer()
+    ..defaultSet.target = organization == SetOrganizationEnum.defaultSet
+        ? defaultSet?.toEntity()
+        : null
+    ..superSet.addAll(superSet.map((e) => e.toEntity()).toList());
 }
 
 @freezed
@@ -47,7 +62,18 @@ class CurrentWorkoutSetType with _$CurrentWorkoutSetType {
   }) = _CurrentWorkoutSetType;
 }
 
+extension ConvertCurrentWorkoutSetType on CurrentWorkoutSetType {
+  SetType toEntity() => SetType()
+    ..exercise.target = exercise.toEntity()
+    ..normalSet.target =
+        type == SetTypeEnum.normalSet ? normalSet?.toEntity() : null;
+}
+
 @freezed
 class CurrentWorkoutNormalSet with _$CurrentWorkoutNormalSet {
   const factory CurrentWorkoutNormalSet() = _CurrentWorkoutNormalSet;
+}
+
+extension ConvertCurrentWorkoutNormalSet on CurrentWorkoutNormalSet {
+  NormalSet toEntity() => NormalSet();
 }
