+++ lib/features/tracker/data/repositories/tracked_workout_repository.dart
@@ -1,6 +1,7 @@
 import 'package:flex_workout_mobile/core/utils/failure.dart';
 import 'package:flex_workout_mobile/db/objectbox.g.dart';
 import 'package:flex_workout_mobile/features/tracker/data/db/tracked_workout_entity.dart';
+import 'package:flex_workout_mobile/features/tracker/data/models/current_workout_model.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/tracked_workout_model.dart';
 import 'package:fpdart/fpdart.dart';
 
@@ -23,22 +24,24 @@ class TrackedWorkoutRepository {
 
   Either<Failure, TrackedWorkoutModel> createTrackedWorkout({
     required String title,
-    required String subtitle,
-    required int durationInMinutes,
-    required DateTime startTimestamp,
-    required List<WorkoutSectionModel> sections,
+    required CurrentWorkout workout,
     String? notes,
   }) {
     try {
+      final now = DateTime.now();
+
+      final totalMinutes = now.difference(workout.startTimestamp).inMinutes;
+
       final workoutToAdd = TrackedWorkout(
         title: title,
-        subtitle: subtitle,
+        subtitle: workout.subtitle,
         notes: notes,
-        durationInMinutes: durationInMinutes,
-        startTimestamp: startTimestamp,
-        updatedAt: DateTime.now(),
-        createdAt: DateTime.now(),
-      )..sections.addAll(sections.map((e) => e.toEntity()).toList());
+        durationInMinutes: totalMinutes,
+        startTimestamp: workout.startTimestamp,
+        updatedAt: now,
+        createdAt: now,
+      )..sections.addAll(workout.sections.map((e) => e.toEntity()).toList());
+
       final id = box.put(workoutToAdd);
 
       final res = box.get(id);
